requirements.txt
# INP214 - Uygulamalı Makine Öğrenmesi
# Python 3.13.4 & Google Colab uyumlu
# pip install -r requirements.txt

numpy>=1.26.0,<2.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
python-pptx
Pillow
scipy>=1.12.0
statsmodels>=0.14.0
yellowbrick>=1.5
ipywidgets>=8.1.0
